src/requirements.txt
google-api==0.1.12
google-api-core==1.22.4
grpcio==1.32.0
grpcio-tools==1.32.0
protobuf==3.13.0
